renamed package name for healthController to package name with expose
